requirements.txt
flask==1.1.4
flask_sqlalchemy==2.5.1
flask_migrate==2.7.0
requests==2.25.1
beautifulsoup4==4.9.3
pandas==1.3.0
psycopg2-binary==2.9.1
apache-airflow==2.1.2
apache-airflow-providers-postgres==2.0.0
apache-airflow-providers-google==6.0.0
SQLAlchemy==1.3.23
python-dotenv==0.19.1